llm-monitor/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Redis for pub/sub
redis==5.0.1
aioredis==2.0.1

# HTTP client for Ollama/backend calls
httpx==0.26.0
aiohttp==3.9.1

# Transformer attention extraction
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0

# Data processing
numpy==1.26.3
pydantic==2.5.3

# Dimensionality reduction for embeddings
scikit-learn==1.3.2
umap-learn==0.5.5
